// libs/my-component/src/lib/Popup/types.ts
export type Placement =
  | 'bottom-right'
  | 'bottom-left'
  | 'bottom-center'
  | 'top-left'
  | 'top-right'
  | 'top-center'
  | 'left-top'
  | 'left-center'
  | 'left-bottom'
  | 'right-top'
  | 'right-center'
  | 'right-bottom';

  export type Position = {
    top: number;
    left: number;
  };